move the use-cases project to separate file (#28)
