+++ src/main/kotlin/logic/usecase/project/DeleteProjectUseCase.kt
@@ -1,4 +1,4 @@
-package org.example.logic.usecase.user
+package org.example.logic.usecase.project
 
 import org.example.logic.repository.ProjectRepository
 import java.util.*
